refactor(adapters): make the tmux backend an extension point
Phase A of the adapter-seam close-out: funnel every tmux invocation through
one overridable spawn primitive so a tmux-family backend (a future native-
Windows psmux) can subclass it without editing tmux_backend.py.

- New adapters/tmux_base.py :: BaseTmuxBackend holds all argv construction and
  every contract method, routed through a single _run() primitive (_tmux uses
  check=True; the ~14 tolerant sites use check=False, keeping their existing
  which-guards and exception handling). TmuxError/TMUX_TIMEOUT_S/
  PARKED_RETURN_DETACH move here.
- tmux_backend.py slims to TmuxMultiplexer(BaseTmuxBackend) plus back-compat
  re-exports (subprocess/shutil/TmuxError/TMUX_TIMEOUT_S/PARKED_RETURN_DETACH).
- Portability guard: TMUX_BACKEND string -> TMUX_BACKENDS set covering both
  files; the sole ["tmux", ...] literal now lives in tmux_base._run.
- Retarget test monkeypatches tmux_backend.{subprocess,shutil} -> tmux_base
  (the spawn seam moved modules).

Behavior-preserving on POSIX. Full suite green (1119 passed); trunk clean.
